+++ primers-java/META-INF/MANIFEST.MF
@@ -8,8 +8,7 @@ Require-Bundle: org.apache.commons.cli,
  biojava,
  org.apache.commons.lang3
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
-Export-Package: org.pooler,
- org.primer3,
+Export-Package: org.primer3,
  org.primer3.dpal,
  org.primer3.libprimer3,
  org.primer3.masker,
@@ -19,5 +18,6 @@ Export-Package: org.pooler,
  org.primer3.primer,
  org.primer3.search.multiplex,
  org.primer3.sequence,
- org.primer3.thal
+ org.primer3.thal,
+ org.ucam.ssb22.pooler
 Import-Package: org.slf4j
