+++ primers-java/src/org/ucam/ssb22/pooler/Amplicons.java
@@ -11,7 +11,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package org.pooler;
+package org.ucam.ssb22.pooler;
 
 import java.io.PrintStream;
 import java.time.Instant;
@@ -21,8 +21,8 @@ import java.util.Comparator;
 import java.util.HashMap;
 import java.util.List;
 
-import org.pooler.Fasta2Bit.SeqData;
-import org.pooler.Fasta2Bit.SeqInfo;
+import org.ucam.ssb22.pooler.Fasta2Bit.SeqData;
+import org.ucam.ssb22.pooler.Fasta2Bit.SeqInfo;
 
 public class Amplicons {
