+++ primers-java/src/org/ucam/ssb22/pooler/PoolerMain.java
@@ -11,9 +11,9 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package org.pooler;
+package org.ucam.ssb22.pooler;
 
-import org.pooler.AllPrimers.PS_cache;
+import org.ucam.ssb22.pooler.AllPrimers.PS_cache;
 
 public class PoolerMain {
 	//	  printf("%s [options] FASTA-file\n",arg0());
