+++ primers-java/src/org/ucam/ssb22/pooler/Splitter.java
@@ -11,7 +11,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package org.pooler;
+package org.ucam.ssb22.pooler;
 
 import java.io.PrintStream;
 import java.time.Instant;
@@ -19,7 +19,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.ThreadLocalRandom;
 
-import org.pooler.AllPrimers.PS_cache;
+import org.ucam.ssb22.pooler.AllPrimers.PS_cache;
 
 public class Splitter {
