+++ primers-java/src/org/ucam/ssb22/pooler/dHS.java
@@ -11,9 +11,13 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package org.pooler.primers;
+package org.ucam.ssb22.pooler;
 
-public class CountResult {
-	public int count = 0;
-	public int tried = 0 ;
+public class dHS {
+	  float dH_kcal_per_mol;
+	  float dS_cal_per_molK;
+	  public dHS(float dH , float dS) {
+		  this.dH_kcal_per_mol = dH;
+		  this.dS_cal_per_molK = dS;
+	  }
 }
