+++ primers-java/src/org/ucam/ssb22/pooler/primers/DegeneratePrimer.java
@@ -11,10 +11,16 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package org.pooler;
+package org.ucam.ssb22.pooler.primers;
 
+import org.ucam.ssb22.pooler.Amplicons.PrimerToFind;
 
-// all operation will be on a 64bit -- Long
-public class BitBasics {
+public abstract class DegeneratePrimer implements IPrimer {
 
+	@Override
+	public final boolean isDegeneratePrimer() {
+		return true;
+	}
+
+	public abstract boolean make2bit(PrimerToFind newPrimerPoss, int possNo, int nPoss);
 }
