+++ primers-java/src/org/ucam/ssb22/pooler/primers/IPrimer.java
@@ -11,7 +11,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package org.pooler.primers;
+package org.ucam.ssb22.pooler.primers;
 
 import java.io.PrintStream;
