+++ primers-java/src/org/ucam/ssb22/pooler/primers/PrimerFactory.java
@@ -11,11 +11,11 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package org.pooler.primers;
+package org.ucam.ssb22.pooler.primers;
 
 import java.io.PrintStream;
 
-import org.pooler.primers64.PrimerFactory64;
+import org.ucam.ssb22.pooler.primers64.PrimerFactory64;
 
 public abstract class PrimerFactory {
