+++ primers-java/src/org/ucam/ssb22/pooler/primers128/Primer128.java
@@ -11,7 +11,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package org.pooler.primers128;
+package org.ucam.ssb22.pooler.primers128;
 
 import java.math.BigInteger;
