+++ primers-java/src/org/ucam/ssb22/pooler/primers64/DegeneratePrimer64.java
@@ -11,15 +11,15 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package org.pooler.primers64;
+package org.ucam.ssb22.pooler.primers64;
 
 import java.io.PrintStream;
 
-import org.pooler.Amplicons.PrimerToFind;
-import org.pooler.HelperMethods;
-import org.pooler.primers.CountResult;
-import org.pooler.primers.DegeneratePrimer;
-import org.pooler.primers.IPrimer;
+import org.ucam.ssb22.pooler.HelperMethods;
+import org.ucam.ssb22.pooler.Amplicons.PrimerToFind;
+import org.ucam.ssb22.pooler.primers.CountResult;
+import org.ucam.ssb22.pooler.primers.DegeneratePrimer;
+import org.ucam.ssb22.pooler.primers.IPrimer;
 
 class DegeneratePrimer64  extends DegeneratePrimer {
 	static String degenerateCombos="TGKCYSBAWRDMHVN";
