+++ primers-java/src/org/ucam/ssb22/pooler/primers64/PrimerFactory64.java
@@ -11,11 +11,11 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package org.pooler.primers64;
+package org.ucam.ssb22.pooler.primers64;
 
-import org.pooler.primers.DegeneratePrimer;
-import org.pooler.primers.IPrimer;
-import org.pooler.primers.PrimerFactory;
+import org.ucam.ssb22.pooler.primers.DegeneratePrimer;
+import org.ucam.ssb22.pooler.primers.IPrimer;
+import org.ucam.ssb22.pooler.primers.PrimerFactory;
 
 /**
  * Current Implementation is 64 bit
